remove some dependencies, geometry.hpp doesn't require sdf_utils.hpp remove exception handling
